Revit-Agent/agent-revit-orchestrator/requirements.txt
flask
pyyaml
transformers
bitsandbytes
torch
spacy